+++ lib/commands/run-test-scenario.ts
@@ -115,6 +115,11 @@ export const builder: yargs.CommandBuilder = {
     string: true,
     demandOption: false,
   },
+  verbose: {
+    describe: "log verbose",
+    default: false,
+    boolean: true,
+  },
 };
 
 export async function handler(argv: yargs.Arguments): Promise<void> {
@@ -178,6 +183,7 @@ export async function handler(argv: yargs.Arguments): Promise<void> {
       from: argv.from,
       to: argv.to,
       runId: argv.runId,
+      verbose: argv.verbose,
     };
     const generator = inversifyGetInstance(PostmanCollectionGenerator, opt);
     await generator.GenerateCollection();